To complete git pages
When I added the html, css, and js files first I didn't expect it to come up with an error to accessing website, and then I have these other files from codesandbox.io I haven't added them to my repository just yet. So I'm hoping this makes a drastic difference and that the website is accessible.